requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers
faiss-cpu
PyPDF2
python-multipart
requests

